requirements.txt
asgiref==3.8.1
astroid==3.3.8
colorama==0.4.6
dill==0.3.9
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
isort==5.13.2
mccabe==0.7.0
platformdirs==4.3.6
PyJWT==2.10.1
pylint==3.3.3
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
tomli==2.2.1
tomlkit==0.13.2
typing_extensions==4.12.2
tzdata==2025.1
